# Add support for connected component traversal & reset